# bondly-api/docs/swagger.yaml
basePath: /
definitions:
...
  models.Proposal:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_time:
        type: string
      id:
        type: integer
      proposer:
        $ref: '#/definitions/models.User'
      proposer_id:
        type: integer
      start_time:
        type: string
      status:
        description: active, passed, rejected, executed
        type: string
      title:
        type: string
      updated_at:
        type: string
      votes_against:
        type: integer
      votes_for:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      custody_wallet_address:
        type: string
      email:
        type: string
      encrypted_private_key:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      nickname:
        type: string
      reputation_score:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      wallet_address:
        type: string
    type: object
...
  response.ResponseAny:
    properties:
      code:
        example: 1000
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  response.ResponseProposal:
    properties:
      code:
        example: 1000
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bondly.com
    name: Bondly API Support
    url: https://bondly.com/support
  description: 这是Bondly项目的API文档，提供用户认证、区块链交互、内容管理和治理功能。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://bondly.com/terms
  title: Bondly API
  version: "1.0"
paths:
...
  /api/v1/proposals:
    get:
      consumes:
      - application/json
      description: 分页获取提案列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取提案列表
      tags:
      - 提案管理
    post:
      consumes:
      - application/json
      description: 创建新的提案
      parameters:
      - description: 提案信息
        in: body
        name: proposal
        required: true
        schema:
          $ref: '#/definitions/models.Proposal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 创建提案
      tags:
      - 提案管理
  /api/v1/proposals/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的提案
      parameters:
      - description: 提案ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 删除提案
      tags:
      - 提案管理
    get:
      consumes:
      - application/json
      description: 根据ID获取提案详情
      parameters:
      - description: 提案ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取提案详情
      tags:
      - 提案管理
    put:
      consumes:
      - application/json
      description: 更新指定提案的信息
      parameters:
      - description: 提案ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的提案信息
        in: body
        name: proposal
        required: true
        schema:
          $ref: '#/definitions/models.Proposal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 更新提案
      tags:
      - 提案管理
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"